Feat: change timestamp to unix and test unSeal encryption data

+++ src/main/kotlin/routes/capsule/types/Request.kt
@@ -23,7 +23,7 @@ data class CreateNewCapsuleRequest(
     val recipients : String,
 
     @RequestInfo(name = "openDate", source = RequestSource.BODY, required = true)
-    val scheduledOpenDate : Long
+    val scheduledOpenDate : Int
 )
